refactor: split github wasm api modules
